requirements.txt
Django==4.0
python-decouple==3.5
djangorestframework==3.13.1
Pillow==9.1.0
joblib==1.1.0
sklearn